Issue with invalid times
Sometimes the Node stopped asking data from the server. I deduct that it is due to errors in the transmission, whereby the calculated time to retransmit could be 0 or negative, leading to it never again asking data from the server, and requiring a reset. Now it is limited to be between 10 and 70 (currently hardcoded, should be a define...).
Errors in the transmission started to be more frequent after installation of the new modem.